backend/requirements.txt
arrow==1.2.2
asgiref==3.5.2
binaryornot==0.4.4
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cookiecutter==2.1.1
Django==4.1
django-browser-reload==1.6.0
django-cache-memoize==0.1.10
django-tailwind==3.4.0
idna==3.3
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-slugify==6.1.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.11
